+++ core/src/test/java/de/fraunhofer/iosb/ilt/faaast/service/assetconnection/fixtures/foo/FooSubscriptionProviderConfig.java
@@ -12,26 +12,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.fraunhofer.iosb.ilt.faaast.service.config.fixtures;
+package de.fraunhofer.iosb.ilt.faaast.service.assetconnection.fixtures.foo;
 
 import de.fraunhofer.iosb.ilt.faaast.service.assetconnection.AssetProviderConfig;
 import de.fraunhofer.iosb.ilt.faaast.service.assetconnection.AssetSubscriptionProviderConfig;
 import de.fraunhofer.iosb.ilt.faaast.service.util.StringHelper;
 import java.util.Objects;
+import org.eclipse.digitaltwin.aas4j.v3.model.builder.ExtendableBuilder;
 
 
-public class DummySubscriptionBasedProviderConfig implements AssetSubscriptionProviderConfig {
+public class FooSubscriptionProviderConfig implements AssetSubscriptionProviderConfig {
 
-    private String nodeId;
+    private String property1;
     private long interval;
 
-    public String getNodeId() {
-        return nodeId;
+    public String getProperty1() {
+        return property1;
     }
 
 
-    public void setNodeId(String nodeId) {
-        this.nodeId = nodeId;
+    public void setProperty1(String property1) {
+        this.property1 = property1;
     }
 
 
@@ -47,7 +48,7 @@ public class DummySubscriptionBasedProviderConfig implements AssetSubscriptionPr
 
     @Override
     public int hashCode() {
-        return Objects.hash(nodeId, interval);
+        return Objects.hash(property1, interval);
     }
 
 
@@ -62,9 +63,9 @@ public class DummySubscriptionBasedProviderConfig implements AssetSubscriptionPr
         if (getClass() != obj.getClass()) {
             return false;
         }
-        final DummySubscriptionBasedProviderConfig other = (DummySubscriptionBasedProviderConfig) obj;
+        final FooSubscriptionProviderConfig other = (FooSubscriptionProviderConfig) obj;
         return Objects.equals(interval, other.interval)
-                && Objects.equals(this.nodeId, other.nodeId);
+                && Objects.equals(this.property1, other.property1);
     }
 
 
@@ -79,9 +80,40 @@ public class DummySubscriptionBasedProviderConfig implements AssetSubscriptionPr
         if (getClass() != other.getClass()) {
             return false;
         }
-        final DummySubscriptionBasedProviderConfig that = (DummySubscriptionBasedProviderConfig) other;
-        return StringHelper.equalsNullOrEmpty(nodeId, that.nodeId)
+        final FooSubscriptionProviderConfig that = (FooSubscriptionProviderConfig) other;
+        return StringHelper.equalsNullOrEmpty(property1, that.property1)
                 && Objects.equals(interval, that.interval);
     }
 
+
+    public static Builder builder() {
+        return new Builder();
+    }
+
+    public static class Builder extends AbstractBuilder<FooSubscriptionProviderConfig, Builder> {
+
+        @Override
+        protected Builder getSelf() {
+            return this;
+        }
+
+
+        @Override
+        protected FooSubscriptionProviderConfig newBuildingInstance() {
+            return new FooSubscriptionProviderConfig();
+        }
+    }
+
+    private abstract static class AbstractBuilder<T extends FooSubscriptionProviderConfig, B extends AbstractBuilder<T, B>> extends ExtendableBuilder<T, B> {
+        public B property1(String value) {
+            getBuildingInstance().setProperty1(value);
+            return getSelf();
+        }
+
+
+        public B interval(long value) {
+            getBuildingInstance().setInterval(value);
+            return getSelf();
+        }
+    }
 }